removed old unused code and refactored storage of everything other than users out of UserService
