+++ src/user/user-router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const AuthService = require('../auth/auth-service');
 const UserService = require('./user-service');
-
+const StorageService = require('../storage/storage-service');
 const userRouter = express.Router();
 const jsonBodyParser = express.json();
 
@@ -14,36 +14,38 @@ userRouter.route('/:user_id/').get(async (req, res, next) => {
   try {
     const db = req.app.get('db');
     const userId = req.user.id;
-    const listIds = await UserService.getListIds(db, userId);
+    const listIds = await StorageService.getListIds(db, userId);
     res.json(listIds);
   } catch (e) {
     next(e);
   }
 });
 
-userRouter.route('/:user_id/:list_id').get(async (req, res, next) => {
-  if (Number(req.params.user_id) !== req.user.id) {
-    return res
-      .status(404)
-      .json({ error: `User has no list with id ${req.params.list_id}` });
-  }
-  try {
-    const db = req.app.get('db');
-    const userId = req.user.id;
-    const requestedListId = Number(req.params.list_id);
-    // check if list belongs to user
-    const listIds = await UserService.getListIds(db, userId);
-    if (!listIds.includes(requestedListId)) {
-      return res.status(401).json(authError);
+userRouter
+  .route('/:user_id/:list_id')
+  .get(async (req, res, next) => {
+    if (Number(req.params.user_id) !== req.user.id) {
+      return res
+        .status(404)
+        .json({ error: `User has no list with id ${req.params.list_id}` });
     }
-    const nodes = await UserService.createStructure(db, requestedListId);
-    nodes.list_id = requestedListId;
-    res.json(nodes);
-    next();
-  } catch (error) {
-    next(error);
-  }
-})
+    try {
+      const db = req.app.get('db');
+      const userId = req.user.id;
+      const requestedListId = Number(req.params.list_id);
+      // check if list belongs to user
+      const listIds = await StorageService.getListIds(db, userId);
+      if (!listIds.includes(requestedListId)) {
+        return res.status(401).json(authError);
+      }
+      const nodes = await StorageService.createStructure(db, requestedListId);
+      nodes.list_id = requestedListId;
+      res.json(nodes);
+      next();
+    } catch (error) {
+      next(error);
+    }
+  })
   .put(jsonBodyParser, async (req, res, next) => {
     if (Number(req.params.user_id) !== req.user.id) {
       return res
@@ -59,10 +61,12 @@ userRouter.route('/:user_id/:list_id').get(async (req, res, next) => {
       }
       const db = req.app.get('db');
 
-      await UserService.insertStructuredList(
+      const listName = req.body.name || 'default';
+
+      await StorageService.insertStructuredList(
         db,
         bookmarksObj,
-        'default',
+        listName,
         req.user.id,
         req.params.list_id
       );
@@ -75,16 +79,14 @@ userRouter.route('/:user_id/:list_id').get(async (req, res, next) => {
 
 userRouter.route('/:user_id/').post(jsonBodyParser, async (req, res, next) => {
   try {
-    // console.log(req.body);
     const bookmarksObj = req.body;
-    // console.log(bookmarksObj);
     if (Object.keys(bookmarksObj).length === 0) {
       return res.status(400).json({ error: 'Empty bookmarks file' });
     }
     const db = req.app.get('db');
     const name = bookmarksObj.name || 'Default';
 
-    const id = await UserService.insertStructuredList(
+    const id = await StorageService.insertStructuredList(
       db,
       bookmarksObj,
       name,
@@ -99,47 +101,6 @@ userRouter.route('/:user_id/').post(jsonBodyParser, async (req, res, next) => {
     next(error);
   }
 });
-
-// userRouter
-//   .route('/:user_id/')
-//   .post(jsonBodyParser, async (req, res, next) => {
-//     const bookmarkArray = [];
-//     try {
-//       const bookmarksObj = req.body;
-//       const bookmarks = bookmarksObj.roots.bookmark_bar.children[0].children;
-//       if (Object.keys(bookmarksObj).length === 0){
-//         return res.status(400).json({error: 'Empty bookmarks file'});
-//       }
-//       const db = req.app.get('db');
-//       await UserService.insertListSimple(db)
-//         .then(list => {
-//           const list_id = list.id;
-//           //for each child of roots, which would be a folder
-//           //but for now assuming just one folder
-//           const folderName = Object.keys(bookmarksObj.roots)[0];
-//           UserService.insertFolderSimple(db, folderName, null)
-//             .then(folder => {
-//               console.log('folder:', folder);
-//               const folder_id = folder[0].id;
-//               UserService.insertListfolderSimple(db, list_id, folder_id)
-//                 .then( () => {
-//                   for (let bookmark of bookmarks){
-//                     const bookmark_id = bookmark.id;
-//                     UserService.insertBookmarkSimple(db, bookmark)
-//                       .then( () => {
-//                         UserService.insertFolderbookmarkSimple(db, folder_id, bookmark_id);
-//                       });
-//                   }
-//                   console.log('bookmarkArray:', bookmarkArray);
-//                   res.status(201).json(bookmarks);
-//                   next();
-//                 });
-//             });
-//         });
-//     } catch(error) {
-//       next(error);
-//     }
-//   });
 
 userRouter.route('/').post(jsonBodyParser, async (req, res, next) => {
   const { password, username, name, email } = req.body;
@@ -196,12 +157,12 @@ userRouter.route('/').post(jsonBodyParser, async (req, res, next) => {
 
 userRouter.route('/').patch(jsonBodyParser, (req, res, next) => {
   const { preview, extra, autosave, color } = req.body;
-  const settings = {preview, extra, autosave, color};
+  const settings = { preview, extra, autosave, color };
   const id = req.user.id;
   console.log(settings);
   console.log(id);
   UserService.patchSettings(req.app.get('db'), id, settings)
-    .then( () => {
+    .then(() => {
       res.status(204).end();
     })
     .catch(next);
@@ -209,9 +170,12 @@ userRouter.route('/').patch(jsonBodyParser, (req, res, next) => {
 
 userRouter.route('/').get(async (req, res, next) => {
   try {
-    const settings = await UserService.getSettings(req.app.get('db'), req.user.id);
-    res.json(settings);  
-  } catch(e) {
+    const settings = await UserService.getSettings(
+      req.app.get('db'),
+      req.user.id
+    );
+    res.json(settings);
+  } catch (e) {
     next(e);
   }
 });
